services/memory-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.3,<3.0.0
pydantic-settings==2.1.0
mem0ai==0.0.11
litellm==1.77.1
slowapi==0.1.9
python-dotenv==1.0.0
python-multipart==0.0.6
typing-extensions>=4.8.0,<5.0.0
aiohttp>=3.8.0,<4.0.0
anyio>=3.6.0,<4.0.0
